Updated with suggestions from marzent: * Moved download funcs to CompatibilityTools * Moved hud funcs to DxvkSettings * Simplified MergeLDPreload * Cleaned up formatting, deleted a commented-out function
